+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to DM-X Video Player will be documented in this file.
 
 - use 'P' instead of 'S' for Parameters window. 'P' can also close the window if opened.
 - remove default video folder settings, save last one used instead.
+- add time signature and bar / beat display (read from .smt cubase tempo track)
 
 ### 🧩 Misc